+++ tower-http/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Breaking changes
 
-None.
+- Add `Cors` for setting [CORS] headers.
 
 # 0.1.1 (July 2, 2021)
 
@@ -37,3 +37,5 @@ None.
 # 0.1.0 (May 27, 2021)
 
 - Initial release.
+
+[CORS]: https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS